fix: avoid useless rerender
